Use new primitives everywhere

+++ src/systems/systems_sdl2.c
@@ -81,17 +81,17 @@ void draw_tilemap(
 	SDL_Texture *tilesheet_texture,
 	const f32v2 camera)
 {
-	const int mapwidth = tilemap->width;
-	const int mapheight = tilemap->height;
-	const int tilewidth = tile_size.w;
-	const int tileheight = tile_size.h;
+	const u32 mapwidth = tilemap->width;
+	const u32 mapheight = tilemap->height;
+	const u32 tilewidth = tile_size.w;
+	const u32 tileheight = tile_size.h;
 
-	for (int l = 0; l < tilemap->num_layers; ++l) {
+	for (usize l = 0; l < tilemap->num_layers; ++l) {
 		const tilemap_layer_t *layer = &tilemap->layers[l];
 
-		for (int y = 0; y < mapheight; ++y) {
-			for (int x = 0; x < mapwidth; ++x) {
-				const int offset = y * mapwidth + x;
+		for (usize y = 0; y < mapheight; ++y) {
+			for (usize x = 0; x < mapwidth; ++x) {
+				const usize offset = y * mapwidth + x;
 				const u8 tile_char = layer->offset_to_char[offset];
 				const tile_enum_t tile_enum = tilemap_encoding->char_to_enum[tile_char];
 				if (tile_enum < TILEMAP_TILE_BEGIN || tile_enum > TILEMAP_TILE_END) continue;
@@ -99,7 +99,7 @@ void draw_tilemap(
 
 				const SDL_Rect srcrect = { tile.x, tile.y, tile.w, tile.h };
 				const SDL_Rect origrect = { x * tilewidth, y * tileheight, tilewidth, tileheight };
-				const SDL_Rect dstrect = { origrect.x - (int)camera.x, origrect.y - (int)camera.y, origrect.w, origrect.h };
+				const SDL_Rect dstrect = { origrect.x - (i32)camera.x, origrect.y - (i32)camera.y, origrect.w, origrect.h };
 				SDL_RenderCopy(renderer, tilesheet_texture, &srcrect, &dstrect);
 			}
 		}
@@ -112,15 +112,15 @@ void draw_tilemap_collision_buffer(
 	SDL_Renderer *renderer,
 	const f32v2 camera)
 {
-	const int mapwidth = tilemap->width;
-	const int mapheight = tilemap->height;
+	const u32 mapwidth = tilemap->width;
+	const u32 mapheight = tilemap->height;
 
 	SDL_SetRenderDrawBlendMode(renderer, SDL_BLENDMODE_BLEND);
 
-	for (int y = 0; y < mapheight; ++y) {
-		for (int x = 0; x < mapwidth; ++x) {
-			const int offset = y * mapwidth + x;
-			const SDL_Rect rect = { x * tile_size.w - (int)camera.x, y * tile_size.h - (int)camera.y, tile_size.w, tile_size.h };
+	for (usize y = 0; y < mapheight; ++y) {
+		for (usize x = 0; x < mapwidth; ++x) {
+			const usize offset = y * mapwidth + x;
+			const SDL_Rect rect = { x * tile_size.w - (i32)camera.x, y * tile_size.h - (i32)camera.y, tile_size.w, tile_size.h };
 			const f32 tile_speed = tilemap->collision_buffer.offset_to_walking_speed[offset];
 			SDL_SetRenderDrawColor(renderer, 255, 0, 0, (f32)100.f * (1.f - tile_speed));
 			SDL_RenderFillRect(renderer, &rect);
